refactored all some issues
