+++ nest_back_end/src/auth/auth.controller.spec.ts
@@ -2,7 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { CreateUserDto } from '../user/dto/create-user.dto';
-import { encryptPassword } from '../util/encrypted.decrypted.util';
+import { encryptPassword } from '../util/password.util';
 
 describe('AuthController', () => {
   let controller: AuthController;
