+++ nest_back_end/src/student/entities/student.entity.ts
@@ -2,21 +2,21 @@ import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
 
 @Entity()
 export class Student {
-    @PrimaryGeneratedColumn()
-    studentId!: number;
-  
-    @Column({nullable: false})
-    studentName!: string;
-  
-    @Column({nullable: false})
-    studentAddress!: string;
-  
-    @Column({nullable: false})
-    studentMobile!: string;
-  
-    @Column({nullable: false})
-    studentDob!: Date;
-  
-    @Column({nullable: false})
-    studentGender!: string;
+  @PrimaryGeneratedColumn()
+  studentId: number;
+
+  @Column({ nullable: false })
+  studentName: string;
+
+  @Column({ nullable: false })
+  studentAddress: string;
+
+  @Column({ nullable: false })
+  studentMobile: string;
+
+  @Column({ nullable: false })
+  studentDob: Date;
+
+  @Column({ nullable: false })
+  studentGender: string;
 }
