+++ nest_back_end/src/user/entities/user.entity.ts
@@ -1,7 +1,7 @@
 import { Entity, Column, PrimaryColumn } from 'typeorm';
 
 @Entity()
-export class User{
+export class User {
   @Column({ nullable: false })
   userName: string;
